% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureSelection.R
\name{annotateData}
\alias{annotateData}
\title{Annotate Data}
\usage{
annotateData(features, tableList, anotpackage, metabList)
}
\arguments{
\item{features}{An ExpressionSet, SummarizedExperiment}

\item{tableList}{A list of comparison tables}

\item{anotpackage}{An annotation package for the corresponding microarray}

\item{metabList}{A list of metabolites for each feature}
}
\value{
Returns an annotated ExpressionSet, SummarizedExperiment or comparison tables
}
\description{
Annotates an ExpressionSet, SummarizedExperiment or comparison table.
}
\author{
Jordi Rofes Herrera
}
