% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchNorm.R
\name{batchNormalization}
\alias{batchNormalization}
\title{Batch Normalization}
\usage{
batchNormalization(
  features,
  method,
  injectionorder,
  batchnum,
  groups,
  qcname,
  covariate,
  covariate2 = NULL
)
}
\arguments{
\item{features}{An ExpressionSet or SummarizedExperiment object}

\item{method}{A string for the normalization method: qcnorm for
injection order and batch normalization and covnorm for covariate normalization}

\item{injectionorder}{A string or numerical indicating the variable in the phenodata
with the injection order or a vector with the injection order.}

\item{batchnum}{A string or numerical indicating the variable in the phenodata
with the batch number}

\item{groups}{A string or numerical indicating the variable in the phenodata
with the grouping variable (QC/SAMPLE/...)}

\item{qcname}{A string with the name of the QC samples in the groups variable}

\item{covariate}{A string or numerical indicating the variable in the phenodata
with the first covariate to normalize}

\item{covariate2}{An optional string or numerical indicating the variable in the phenodata
with the second covariate to normalize}
}
\value{
Returns a covariate/batch corrected ExpressionSet or SummarizedExperiment
}
\description{
Makes the batch normalization for injection order (Metabolomics)
using the pmp package and covariates (Transcriptomics) using the pvca package.
}
\author{
Jordi Rofes Herrera
}
