% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupComp.R
\name{compPlot}
\alias{compPlot}
\title{Feature comparison plot}
\usage{
compPlot(features, subset, groupvar, ...)
}
\arguments{
\item{features}{A SummarizedExperiment or ExpressionSet}

\item{subset}{A numeric vector indicating a subset of features to plot}

\item{groupvar}{A numeric or string indicating the variable from the phenodata
to compare}
}
\value{
A ggplot2 object
}
\description{
Creates a box plot for each selected features to compare groups.
}
\author{
Jordi Rofes Herrera
}
