% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchNorm.R
\name{featureBatchPVCA}
\alias{featureBatchPVCA}
\title{PVCA batch effect plot}
\usage{
featureBatchPVCA(features, phenovars, threshold)
}
\arguments{
\item{features}{An ExpressionSet object}

\item{threshold}{A precentile value for the minimum variance the selected
principal components need to explain}

\item{phenovar}{A numeric or string vector indicating the columns of the pehnodata to include
in the pvca analysis}
}
\value{
Returns a ggplot object or a ggplotly
}
\description{
Creates a column plot with the variance of each covariate in the
principal variation component analysis
}
\author{
Jordi Rofes Herrera
}
