% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchNorm.R
\name{featurebatchQc}
\alias{featurebatchQc}
\title{Feature sum intensity plot}
\usage{
featurebatchQc(
  features,
  groupvar,
  injectionvar,
  logscale = FALSE,
  interactive = TRUE
)
}
\arguments{
\item{features}{A SummarizedExperiment with features}

\item{groupvar}{A numeric indicating the column to use as grouping factor or a
character indicating it's name}

\item{injectionvar}{An optional string or numeric indicating the variable
in the phenodata used to indicate the injection order, no value will default
to the file order}

\item{logscale}{log10 intensity transformation}

\item{interactive}{A boolean indicating if the plot will be converted to an
interactive `ggplotly()`}
}
\value{
Returns a ggplot object or a ggplotly
}
\description{
Creates a scatter plot with the sum of intensities of all features
for each sample ordered by injection order to visualize batch effect
}
\author{
Jordi Rofes Herrera
}
