% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureSelection.R
\name{featureSign}
\alias{featureSign}
\title{Feature Sign}
\usage{
featureSign(features, groupvar, boot)
}
\arguments{
\item{features}{An ExpressionSet or SummarizedExperiment object}

\item{groupvar}{A string or numeric indicating the group variable in the phenodata for classification (two factor only).}
}
\value{
Returns a biosign object
}
\description{
Uses the biosign function from the biosigner package to create a biosign object for the given features
}
\author{
Jordi Rofes Herrera
}
