% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureSelection.R
\name{featureVarPlot}
\alias{featureVarPlot}
\title{Variance Plot}
\usage{
featureVarPlot(features, varfun, interactive = TRUE)
}
\arguments{
\item{features}{A SummarizedExperiment or ExpressionSet}

\item{varfun}{A function to calculate for each feature, usually IQR / SD or function(x){sd(x)/mean(x)}}

\item{interactive}{A boolean indicating if the plot will be converted to an
interactive `ggplotly()`}
}
\value{
Returns a ggplot object or a ggplotly
}
\description{
Creates a scatter plot with the ordered feature values after
evaluating the given function for each feature usually used to visualize the
distribution of IQR, standard deviation or coefficient of variation (CV) for each feature.
}
\author{
Jordi Rofes Herrera
}
