% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureSelection.R
\name{geneFeatureFilter}
\alias{geneFeatureFilter}
\title{Gene Feature Filter}
\usage{
geneFeatureFilter(
  features,
  entrez,
  rem.dupEntrez,
  varfilt,
  varcutoff,
  var.func
)
}
\value{
Returns filtered gene features
}
\description{
Wrapper for the nsFilter() function. Use ?nsFilter for param information
}
\author{
Jordi Rofes Herrera
}
