% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcrProc.R
\name{ggDensityPlot}
\alias{ggDensityPlot}
\title{Microarray intensity density plot}
\usage{
ggDensityPlot(
  object,
  groupvar,
  interactive = TRUE,
  nsamp = 10000,
  ytrans = NA,
  xtrans = "log2"
)
}
\arguments{
\item{object}{A microarray data in a ExpressionSet or GeneFeatureSet}

\item{groupvar}{A numeric indicating the column to use as grouping factor or a
character indicating it's name}

\item{interactive}{A boolean indicating if the plot will be converted to an
interactive `ggplotly()`}

\item{nsamp}{The number of samples used in the plotting to reduce computation
time}

\item{xtrans}{A transformation applied to the represented intensities, usually
log2 for microarray analysis}
}
\value{
Returns a ggplot object or a ggplotly
}
\description{
Creates a density distribution plot from a microarray object with
 `exprs()` such as an ExpressionSet
}
\examples{
plot_crayons()
}
\author{
Jordi Rofes Herrera
}
