% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcrProc.R
\name{ggExprDistrPlot}
\alias{ggExprDistrPlot}
\title{Microarray intensity distribution boxplot/violin plot}
\usage{
ggExprDistrPlot(
  object,
  violin = TRUE,
  groupvar,
  interactive = TRUE,
  nsamp = 1e+05,
  ytrans = "log2"
)
}
\arguments{
\item{object}{A microarray data in a ExpressionSet or GeneFeatureSet}

\item{violin}{A boolean indicating if the plot will be a violin plot/boxplot}

\item{groupvar}{A numeric indicating the column to use as grouping factor or a
character indicating it's name}

\item{interactive}{A boolean indicating if the plot will be converted to an
interactive `ggplotly()`}

\item{nsamp}{The number of samples used in the plotting to reduce computation
time}

\item{ytrans}{A transformation applied to the represented intensities, usually
log2 for microarray analysis}
}
\value{
Returns a ggplot object or a ggplotly
}
\description{
Creates a boxplot/violin plot from a microarray object with
 `exprs()` such as an ExpressionSet
}
\author{
Jordi Rofes Herrera
}
