% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcrProc.R
\name{ggPCAplot}
\alias{ggPCAplot}
\title{Microarray sample PCA plot}
\usage{
ggPCAplot(object, pc = c(1, 2), groupvar, scale = TRUE, interactive = TRUE)
}
\arguments{
\item{object}{Microarray data from a ExpressionSet or GeneFeatureSet}

\item{pc}{A vector of length two indicanting the principal components to plot
on each axis}

\item{groupvar}{A numeric indicating the column to use as grouping factor or a
character indicating it's name}

\item{scale}{A boolan indicating if the PCA will be scaled, (the default is
TRUE).}

\item{interactive}{A boolean indicating if the plot will be converted to an
interactive `ggplotly()`}
}
\value{
Returns a ggplot object or a ggplotly
}
\description{
Creates a PCA plot from a microarray object with `exprs()` such
 as an ExpressionSet
}
\examples{
plot_crayons()
}
\author{
Jordi Rofes Herrera
}
