% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaboVis.R
\name{ggTicQuality}
\alias{ggTicQuality}
\title{Total ion count plot}
\usage{
ggTicQuality(
  object,
  filenum = NA,
  pheno_var = 2,
  pheno_filter,
  logscale = TRUE,
  violin = FALSE,
  interactive = TRUE,
  order = FALSE
)
}
\arguments{
\item{object}{A OnDiskXCMSnExp or XCMSnExp object}

\item{filenum}{An optional numeric vector indicating the files to plot}

\item{pheno_var}{A numeric or string indicating the variable from the phenodata
to use as file groups.}

\item{pheno_filter}{An optiona string of a variable name in pheno_var to filter from plotting}

\item{logscale}{A boolean to apply a log10 transformation of the intensities}

\item{interactive}{A boolean indicating if the plot will be converted to a plotly}

\item{order}{A boolean indicating if the file order should be used for the order of the samples}

\item{violing}{A boolean indicating if a violin plot should be plot instead of a boxplot}
}
\value{
A ggplot2 or plotly violin/boxplot TIC plot
}
\description{
Creates a violin/boxplot TIC plot from an XCMS object
}
\author{
Jordi Rofes Herrera
}
