% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupComp.R
\name{groupComp}
\alias{groupComp}
\title{Basic group comparison}
\usage{
groupComp(
  features,
  groupvar,
  elimlab = "none",
  test,
  adj.method = "fdr",
  paired,
  var.equal = FALSE
)
}
\arguments{
\item{features}{An ExpressionSet or SummarizedExperiment object}

\item{groupvar}{A numeric or string indicating the variable from the phenodata
from which the comparisons will be made}

\item{elimlab}{An optional string of a variable name in groupvar indicating a
group that will not be used for comparisons.}

\item{test}{A string indicating the test to be perform: t-test or wilcox}

\item{adj.method}{A string indicating the p-value adjust methodology for multiple testing}

\item{paired}{A boolean indicating if the t-test should be paired or not}

\item{var.equal}{A boolean indicating if the variance in the t-test are considered equal or not}
}
\value{
Returns a list of comparison tables for each pair of groups
}
\description{
Compares different groups of features using log2(Fold-Change) and using
the p-value from a t-test or wilcox test
}
\author{
Jordi Rofes Herrera
}
