% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureSelection.R
\name{groupFeatureComp}
\alias{groupFeatureComp}
\title{Limma group comparison}
\usage{
groupFeatureComp(features, modelMatrix, contrastMat, adjpval = "fdr")
}
\arguments{
\item{features}{An ExpressionSet object}

\item{modelMatrix}{A model matrix}

\item{contrastMat}{A contrast matrix}

\item{adjpval}{A string indicating the p-value adjust methodology "fdr" or "bonferroni"
to create the model matrix with}
}
\value{
A list of comparisson tables for each pair of groups
}
\description{
Calculates comparison tables with the Limma package for each group pair
}
\author{
Jordi Rofes Herrera
}
