% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureSelection.R
\name{groupFeatureVolcano}
\alias{groupFeatureVolcano}
\title{Volcano plot}
\usage{
groupFeatureVolcano(
  comptable,
  adj.pvalue = TRUE,
  interactive = TRUE,
  jitterseed = 123
)
}
\arguments{
\item{comptable}{A comparison table}

\item{adj.pvalue}{Boolean indicating if the adjusted p-values will be used}

\item{interactive}{Boolean indicating if a plotly will be output instead of a ggplot}

\item{jitterseed}{A seed for the random jitter to avoid names overlaps}
}
\value{
Returns filtered gene features
}
\description{
Creates a volcano plot for a comparison table
}
\author{
Jordi Rofes Herrera
}
