% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupComp.R
\name{heatmapPlot}
\alias{heatmapPlot}
\title{Heatmap plot}
\usage{
heatmapPlot(features, groupvar)
}
\arguments{
\item{features}{A SummarizedExperiment or ExpressionSet}

\item{groupvar}{A numeric or string indicating the variable from the phenodata
to use as sample names.}
}
\value{
A heatmap plot
}
\description{
Creates a heatmap plot with clustering in both axis
}
\author{
Jordi Rofes Herrera
}
