% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcrProc.R
\name{importRawTranscript}
\alias{importRawTranscript}
\title{Wrapper function for .idat, .cel and geoDatasets importing}
\usage{
importRawTranscript(datapath, sampledata, geoid, header = TRUE, sep = ",")
}
\arguments{
\item{datapath}{A string with the directory where the files are located or a vector with all the files to import}

\item{sampledata}{A string indicating the phenodata location as a comma separated values or excel file.}

\item{geoid}{A string indicating a geoDataset accesion entry}

\item{header}{A boolean indicating if the sampledata has a header with the variable names}

\item{sep}{A string indicating the separator of the sampledata file if it's a .csv file, usually a space or a comma.}

\item{groupvar}{A numeric indicating the column to use as grouping factor or a character indicating it's name}
}
\value{
Returns an object with the expression data.
}
\description{
Imports .idat, .cel files from microarray expression data into their respective formats and the download of geoDatasets.
To import illumina datasets it is required the appropiate annotation package
}
\examples{
plot_crayons()
}
\author{
Jordi Rofes Herrera
}
