% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupComp.R
\name{loadingsPlot}
\alias{loadingsPlot}
\title{Loadings Plot}
\usage{
loadingsPlot(pcdt, pc = 1, interactive = TRUE, len = 100)
}
\arguments{
\item{pcdt}{The prcomp object to use for plotting}

\item{pc}{A numeric indicating the principal component selected}

\item{interactive}{Boolean indicating if a plotly will be output instead of a ggplot}
}
\value{
A ggplot2 or plotly dot plot with the loadings of selected principal component
}
\description{
Creates dot chart with the loadings of selected principal component.
}
\author{
Jordi Rofes Herrera
}
