% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureSelection.R
\name{metabFeatureFilter}
\alias{metabFeatureFilter}
\title{Metablomics feature filter}
\usage{
metabFeatureFilter(
  features,
  groupvar,
  blankfilt = FALSE,
  blankFoldChange = 2,
  blankname = "blank",
  samplename,
  cvqcfilt = FALSE,
  cvqc_thr = 30,
  qcname = "QC",
  nafilter = FALSE,
  naratioThr,
  naratioMethod,
  varfilter = FALSE,
  varfun,
  varthr,
  varquant,
  intfilter = FALSE,
  intensitythr,
  ism0 = FALSE,
  hasan = FALSE,
  sampfilter = FALSE,
  maxmv,
  filtername,
  prepro,
  preprofuns,
  mvimpmethod
)
}
\arguments{
\item{features}{A SummarizedExperiment}

\item{groupvar}{A string or numeric indicating the variable in the phenodata
with the group variable (sample/qc/blank)}

\item{blankfilt}{Boolean indicating blank filtering}

\item{blankFoldChange}{A numeric indicating the fold change to filter features with high intensity in blank}

\item{blankname}{A string indicating the blank name in the groupvar}

\item{samplename}{A string indicating the sample name in the groupvar}

\item{cvqcfilt}{A boolean indicating the filter of features with high RDS in QC samples}

\item{cvqc_thr}{A numeric indicating the maximum percentage of RDS in QC samples}

\item{qcname}{A string indicating the qcname in the groupvar variable}

\item{nafilter}{A boolean indicating the filter of features with high percentage of missing values}

\item{naratioThr}{A numeric indicating the fraction of NA values allowed in the features}

\item{naratioMethod}{A string indicating the method for calculating the NA ratio: can be "QC" for within QC samples,
"within" for within each sample class in groupvar and "across" for across all samples.}

\item{varfilter}{A boolean indicating filtering depending on variability criteria}

\item{varfun}{A function to calculate for each feature, usually IQR / SD or function(x){sd(x)/mean(x)}}

\item{varthr}{The minimum threshold of variability to keep the features}

\item{varquant}{A boolean indicating if the threshold is given as a precentile or an absolute value (TRUE/FALSE)}

\item{intfilter}{A boolean indicating to filter by sample intensity}

\item{intensitythr}{A value indicating the minimum intensity of atleast one sample in each feature}

\item{ism0}{A boolean indicating if filtering of features with non M0 annotations
(Requires annotated SummarizedExperiment as done with the pre-processing functions)}

\item{hasan}{A boolean indicating to filter features with no annotation
(Requires annotated SummarizedExperiment as done with the pre-processing functions)}

\item{maxmv}{A numeric indicating the fraction of maximum missing values in each sample}

\item{filtername}{An optional name indicating the name of a variable in groupvar to eliminate all it's samples.
Useful to eliminate QC samples after using them for all pre-processing.}

\item{prepro}{A boolean indicating to do a pre-processing of all features}

\item{preprofuns}{A vector of strings indicating all the functions applied to the features. Available options are:
"pqn" point quotien normalization, "sum" normalization to the sum and "mvImp" multivariate imputation.}

\item{mvimpmethod}{A string indicating the multivariate imputation method: "knn" or "rf".}

\item{samplfilter}{A boolean indicating to filter samples with high missing values}
}
\value{
Returns a SummarizedExperiment with filtered features
}
\description{
Filters metabolomic features with a variety of criteria using different functions
}
\author{
Jordi Rofes Herrera
}
