% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaboProc.R
\name{metaboImport}
\alias{metaboImport}
\title{Import metabolomics files .mzML/.mzXML}
\usage{
metaboImport(filedir, phenodata, injectionvar)
}
\arguments{
\item{phenodata}{A NAnnotatedDataFrame with the phenodata.}

\item{injectionvar}{An optional string or numeric indicating the variable from
the phenodata to use as the injection order. If not given, the order will be the
file order.}

\item{file}{A string indicating the phenodata path.}
}
\value{
A OnDiskXCMSnExp object
}
\description{
Imports .mzML/.mzXML
}
\author{
Jordi Rofes Herrera
}
