% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaboProc.R
\name{metaboProc}
\alias{metaboProc}
\title{Metabolomics files process}
\usage{
metaboProc(
  object,
  polarity = "positive",
  groupvar,
  peakwidth,
  noise,
  snthresh,
  ppm,
  expandrt,
  binsize,
  minFraction,
  bw,
  annotation = "camera",
  cliqsamp = NA,
  mergepeaks = TRUE,
  rtadjust = TRUE,
  group = TRUE,
  fill = TRUE
)
}
\arguments{
\item{object}{A OnDiskMSnExp/MSnExp object}

\item{polarity}{A string indicating the polarity of the experiment "positive" or "negative"}

\item{groupvar}{A string or numeric indicating the group variable used to group features}

\item{peakwidth}{A two length numeric vector indicating the width of a peak}

\item{noise}{A numeric indicating the maximum noise of a chromatographic}

\item{snthresh}{A numeric indicating the minimum signal to noise ratio of a chromatographic peak}

\item{ppm}{A numeric indicating the ppm error of the instrument}

\item{expandrt}{A numeric indicating the RT seconds to search for overlapping peaks}

\item{binsize}{A numeric indicating the binsize to perform RT adjusting}

\item{minFraction}{A numeric indicating the minimum fraction of peaks to group features}

\item{bw}{A numeric indicating the binwidth to group features.}

\item{annotation}{A string indicating the annotation methodology. One of: "cliqueMS", "camera" or "none".}

\item{cliqsamp}{A numeric indicating the sample from wich the cliqueMS annotation will be perform.
Usually a pool QC sample.}

\item{mergepeaks}{A boolean indicating if peak merging will be perform}

\item{rtadjust}{A boolean indicating if retention time adjustment will be perform}

\item{group}{A boolean indicating if peak grouping will be perform}

\item{fill}{A boolean indicating if peak filling will be perform}
}
\value{
A processed XCMSnExp object, a annotation object if annotation was selected
and a SummarizedExperiment with feature information.
}
\description{
Processes a OnDiskMSnExp/MSnExp object with XCMS and annotates it with CAMERA or CliqueMS.
}
\author{
Jordi Rofes Herrera
}
