% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machineLearning.R
\name{mlFit}
\alias{mlFit}
\title{Caret model fitting}
\usage{
mlFit(
  dt,
  groupvar,
  method,
  metric = "Accuracy",
  cvmethod = "loo",
  kfolds = 10,
  ntimes = 5,
  preproc = c("center", "scale"),
  tlength = 20,
  ...
)
}
\arguments{
\item{dt}{A SummarizedExperiment or ExpressionSet object}

\item{groupvar}{A numeric or string indicating the variable from the phenodata
to use for grouping (must be a factor of length two)..}

\item{metric}{A string with the metric to compute for each training step (see ?train for more details)}

\item{cvmethod}{A string indicating de crossvalidation method "loo" for leave one out.}

\item{ntimes}{Number of partition to create}

\item{preproc}{A string vector with pre-processing options. See ?train for more details.}

\item{filenum}{An optional numeric vector indicating the files to plot}

\item{kfold}{Number of folds for k-fold crossvalidation}

\item{tlenght}{The granularity of the training parameters}
}
\value{
A train object
}
\description{
Trains selected machine learning model from features
}
\author{
Jordi Rofes Herrera
}
