% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machineLearning.R
\name{mlPredictCM}
\alias{mlPredictCM}
\title{Caret confusion matrix}
\usage{
mlPredictCM(mlmod, newdt, prepro_obj, groupvar, posclass)
}
\arguments{
\item{mlmod}{A machine learning model with a predict method.}

\item{newdt}{A SummarizedExperiment or ExpressionSet object with the test data}

\item{prepro_obj}{An optional pre-processing object created by the preProcess() function. Not required with a training object.}

\item{groupvar}{A numeric or string indicating the variable from the phenodata
to use for grouping (must be a factor of length two).}

\item{posclass}{A string indicating the positive class from the groupvar}
}
\value{
A confusion matrix
}
\description{
Calculates the caret confusion matrix with the machine learning model and new data.
}
\author{
Jordi Rofes Herrera
}
