% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machineLearning.R
\name{mlPredictROC}
\alias{mlPredictROC}
\title{Receiver Operand Curve (ROC)}
\usage{
mlPredictROC(mlmod, newdt, prepro_obj, groupvar, posclass)
}
\arguments{
\item{mlmod}{A machine learning model with a predict method with "prob" option.}

\item{newdt}{A SummarizedExperiment or ExpressionSet object with the test data}

\item{groupvar}{A numeric or string indicating the variable from the phenodata
to use for grouping (must be a factor of length two).}

\item{posclass}{A string indicating the positive class from the groupvar}
}
\value{
A ROC curve plot
}
\description{
Calculates the ROC curve with the machine learning model and new data.
}
\author{
Jordi Rofes Herrera
}
