% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureSelection.R
\name{model.mat}
\alias{model.mat}
\title{Model matrix}
\usage{
model.mat(features, phenovar)
}
\arguments{
\item{features}{An ExpressionSet object}

\item{phenovar}{A string or numeric indicating the group variable in the phenodata
to create the model matrix with}
}
\value{
Returns filtered gene features
}
\description{
Creates a model matrix from an ExpressionSet
}
\author{
Jordi Rofes Herrera
}
