% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupComp.R
\name{multipca}
\alias{multipca}
\title{Principal component analysis}
\usage{
multipca(features, prefuns, method, groupvar, qcname, scale = TRUE, ...)
}
\arguments{
\item{features}{An ExpressionSet or SummarizedExperiment object}

\item{prefuns}{An option string vector indicating pre-processing functions applied
to the features before the PCA. Recommended for metabolomics data to impute missing values.
Pre-processing functions available are "pqn" normalization, "sum" normalization, "glog"
transformation and "mvImp" multivariate imputation}

\item{method}{A string indicating the multivariate imputation method. Usually "knn" or "rf"}

\item{groupvar}{A numeric or string indicating the variable from the phenodata with the QC
samples to apply the pre-processing functions that require QC.}

\item{qcname}{The QC name from the groupvar}

\item{scale}{Boolean indicating if the pca should be scaled or not.}
}
\value{
A prcomp object
}
\description{
Calculates the PCA for a SummarizedExperiment or ExpressionSet
}
\author{
Jordi Rofes Herrera
}
