% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupComp.R
\name{pcaPlot}
\alias{pcaPlot}
\title{Principal component analysis plot}
\usage{
pcaPlot(
  pcdt,
  object,
  pc = c(1, 2),
  groupvar,
  interactive = TRUE,
  ellipse = TRUE,
  ellipse.type = "t",
  level = 0.95,
  bygroup = TRUE
)
}
\arguments{
\item{pcdt}{The prcomp object to use for plotting}

\item{object}{The ExpressionSet or SummarizedExperiment object used when creating the PCA}

\item{pc}{A length two vector with the principal components to be plot in each axis (x, y).}

\item{groupvar}{A numeric or string indicating the variable from the phenodata with the QC
samples to apply the pre-processing functions that require QC.}

\item{interactive}{Boolean indicating if a plotly will be output instead of a ggplot}

\item{ellipse}{A boolean indicating if a ellipse should be drawn}

\item{ellipse.type}{The type of ellipse drawn. See ?stat_ellipse for more information}

\item{level}{The confidence level of the ellipse}

\item{bygroup}{A boolean indicating if ellipses should be drawn for each group. (If there are enough points)}
}
\value{
A ggplot2 or plotly pca scatter plot
}
\description{
Creates a scatter plot with the selected principal components
}
\author{
Jordi Rofes Herrera
}
