% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaboProc.R
\name{phenoImport}
\alias{phenoImport}
\title{Import phenodata}
\usage{
phenoImport(file, header, sep)
}
\arguments{
\item{file}{A string indicating the phenodata path.}

\item{header}{A boolean indicating if the phenodata has a header with the
variable names}

\item{sep}{A string indicating the separator of the phenodata file, if it's a
.csv file, usually a space or a comma.}

\item{geoid}{A string indicating a geoDataset accesion entry}
}
\value{
Returns a data frame with the imported phenodata
}
\description{
Imports .csv/.xlsx/.xls/.tsv/.txt files
}
\author{
Jordi Rofes Herrera
}
