% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcrProc.R
\name{procTranscript}
\alias{procTranscript}
\title{Processing function for illumina and affymetrix microarrays illumina
 expression normalization and RMA for affymetrix.}
\usage{
procTranscript(object, method, annotationTable)
}
\arguments{
\item{object}{A microarray data in a ExpressionSet or GeneFeatureSet}

\item{method}{String specifying the illumina normalization method methods are:
"quantile", "qspline", "vsn", "rankInvariant", "median" and "none".}
}
\value{
Returns a processed ExpressionSetIllumina for illumina data or an
 ExpressionSet
}
\description{
Normalizes a GeneFeatureSet using the RMA method or an
 ExpressionSetIllumina using the selected methodology
}
\author{
Jordi Rofes Herrera
}
